Adjust layout for Safari
