+++ src/containers/SldEditor/sldEditor.css
@@ -128,12 +128,16 @@
   justify-content: center;
   align-items: center;
   padding: 20px 0 20px 0;
+  min-height: 170px;
+  height: 20%;
 }
 .sld-item > a {
-  width: 80%;
-  padding-bottom: 55%;
+  height: 100%;
+  padding-left: 56.25%;
   position: relative;
   max-width: 330px;
+  min-width: 120px;
+  flex-grow: 1;
 }
 
 .sld {
@@ -148,6 +152,7 @@
   bottom: 0;
   top: 0;
   max-width: 330px;
+  height: 100%;
 }
 
 .sld-item-content {
@@ -585,8 +590,8 @@
   }
   .sld-item > a {
     padding-bottom: 0;
-    width: 80%;
-    height: 90%;
+    width: 100%;
+    height: 190px;
   }
 
   .middle-editor > .edit-panel {
@@ -618,6 +623,8 @@
   .sld-item-title {
     font-size: 1.3rem;
     min-height: 230px;
+    margin: 0 18px;
+    width: 15%;
   }
 
   .sld-item {
@@ -631,6 +638,7 @@
   .sld-item-tools {
     height: 210px;
     margin: 0 18px;
+    width: 15%;
   }
 
   .hide-tool {
@@ -654,6 +662,11 @@
 
   #add-sld-btn {
     margin: 25px auto 80px;
+  }
+
+  .control-panel {
+    overflow-x: hidden;
+    overflow-y: scroll;
   }
 
   .mobile-control-panel,
@@ -736,11 +749,5 @@
   .sld-copy-btn:hover,
   .sld-item-title > .trash-bin:hover {
     transform: none;
-  }
-}
-
-@media (max-width: 400px) {
-  .sld-item > a {
-    padding-bottom: 55%;
   }
 }
